use new hook to fetch activity and reflections
